Changes migration and adds line to seeds.

+++ db/migrate/20230915164810_delete_users.rb
@@ -1,5 +1,7 @@
 class DeleteUsers < ActiveRecord::Migration[7.0]
   def change
-    drop_table :users
+    if table_exists?(:users)
+      drop_table :users
+    end
   end
 end
